moving things around, adding experimental config for dom-fb
