+++ packages/react-server/src/forks/ReactServerStreamConfig.dom-fb-experimental.js
@@ -9,4 +9,6 @@
 
 export * from '../ReactServerStreamConfigFB';
 
-export function scheduleWork(callback: () => void) {}
+export function scheduleWork(callback: () => void) {
+  callback();
+}
